requirements.txt
# Be Observant Control Plane Dependencies

# FastAPI and server
fastapi==0.129.0
starlette==0.49.1
uvicorn[standard]==0.27.1
uvloop==0.19.0

# HTTP client
httpx==0.27.0

# Data validation
pydantic==2.12.5
pydantic-settings==2.11.0

# Additional utilities
python-multipart==0.0.22
cryptography==46.0.5
aiokafka==0.8.0
pyotp==2.8.0
protobuf>=4.22.0
opentelemetry-proto>=1.15.0
psutil==5.9.8

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
bcrypt==3.2.0
PyJWT[crypto]==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiosmtplib==1.1.6
tenacity==8.2.2
PyYAML==6.0.2
redis==5.0.8
hvac